Performance improvements.
