+++ vaadin-demo-micro-service/src/main/java/de/muenchen/demo/service/ApplicationStartup.java
@@ -1,6 +1,6 @@
 package de.muenchen.demo.service;
 
-import de.muenchen.auditing.AuditingService;
+import de.muenchen.auditing.AuditingServiceProducer;
 import de.muenchen.demo.service.domain.AuditingUserRepositoryImpl;
 import de.muenchen.eventbus.EventBus;
 import org.hibernate.search.jpa.FullTextEntityManager;
@@ -15,6 +15,7 @@ import org.springframework.stereotype.Component;
 import javax.persistence.EntityManager;
 
 @Component
+//@ComponentScan(basePackages = {"de.muenchen.auditing"})
 public class ApplicationStartup implements ApplicationListener<ContextRefreshedEvent> {
 
 	 private static final Logger LOG = LoggerFactory.getLogger(ApplicationStartup.class);
@@ -35,8 +36,8 @@ public class ApplicationStartup implements ApplicationListener<ContextRefreshedE
 	@Override
 	public void onApplicationEvent(final ContextRefreshedEvent event) {
 
-		// Initialize AuditingService
-		new AuditingService(entityManager.getEntityManagerFactory(), eventBus, repo).init();
+		// Initialize AuditingServiceProducer
+		new AuditingServiceProducer(entityManager.getEntityManagerFactory(), eventBus, repo).init();
 
 		FullTextEntityManager fullTextEntityManager = Search.getFullTextEntityManager(entityManager);
 		try {
